+++ src/Unbound/Generics/LocallyNameless/Subst.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeOperators #-}
+{-# OPTIONS_GHC -fexpose-all-unfoldings #-}
 
 -- |
 -- Module     : Unbound.Generics.LocallyNameless.Subst
@@ -75,13 +76,15 @@ class Subst b a where
   -- | This is the only method that must be implemented
   isvar :: a -> Maybe (SubstName a b)
   isvar _ = Nothing
-
+  {-# INLINE isvar #-}
+  
   -- | This is an alternative version to 'isvar', useable in the case
   --   that the substituted argument doesn't have *exactly* the same type
   --   as the term it should be substituted into.
   --   The default implementation always returns 'Nothing'.
   isCoerceVar :: a -> Maybe (SubstCoerce a b)
   isCoerceVar _ = Nothing
+  {-# INLINE isCoerceVar #-}
 
   -- | @'subst' nm e tm@ substitutes @e@ for @nm@ in @tm@.  It has
   -- a default generic implementation in terms of @isvar@
@@ -95,6 +98,8 @@ class Subst b a where
           Just (SubstCoerce m f) | m == n -> fromMaybe x (f u)
           _ -> to $ gsubst n u (from x)
       else error $ "Cannot substitute for bound variable " ++ show n
+  {-# INLINE subst #-}
+
 
   substs :: [(Name b, b)] -> a -> a
   default substs :: (Generic a, GSubst b (Rep a)) => [(Name b, b)] -> a -> a
@@ -107,6 +112,7 @@ class Subst b a where
           _ -> to $ gsubsts ss (from x)
     | otherwise =
       error $ "Cannot substitute for bound variable in: " ++ show (map fst ss)
+  {-# INLINE substs #-}
 
 ---- generic structural substitution.
 class GSubst b f where
@@ -116,22 +122,32 @@ class GSubst b f where
 instance Subst b c => GSubst b (K1 i c) where
   gsubst nm val = K1 . subst nm val . unK1
   gsubsts ss = K1 . substs ss . unK1
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 instance GSubst b f => GSubst b (M1 i c f) where
   gsubst nm val = M1 . gsubst nm val . unM1
   gsubsts ss = M1 . gsubsts ss . unM1
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 instance GSubst b U1 where
   gsubst _nm _val _ = U1
   gsubsts _ss _ = U1
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 instance GSubst b V1 where
   gsubst _nm _val = id
   gsubsts _ss = id
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 instance (GSubst b f, GSubst b g) => GSubst b (f :*: g) where
   gsubst nm val (f :*: g) = gsubst nm val f :*: gsubst nm val g
   gsubsts ss (f :*: g) = gsubsts ss f :*: gsubsts ss g
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 instance (GSubst b f, GSubst b g) => GSubst b (f :+: g) where
   gsubst nm val (L1 f) = L1 $ gsubst nm val f
@@ -139,21 +155,47 @@ instance (GSubst b f, GSubst b g) => GSubst b (f :+: g) where
 
   gsubsts ss (L1 f) = L1 $ gsubsts ss f
   gsubsts ss (R1 g) = R1 $ gsubsts ss g
+  {-# INLINE gsubst #-}
+  {-# INLINE gsubsts #-}
 
 -- these have a Generic instance, but
 -- it's self-refential (ie: Rep Int = D1 (C1 (S1 (Rec0 Int))))
 -- so our structural GSubst instances get stuck in an infinite loop.
-instance Subst b Int where subst _ _ = id; substs _ = id
+instance Subst b Int where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
-instance Subst b Bool where subst _ _ = id; substs _ = id
+instance Subst b Bool where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
-instance Subst b () where subst _ _ = id; substs _ = id
+instance Subst b () where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
-instance Subst b Char where subst _ _ = id; substs _ = id
+instance Subst b Char where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
-instance Subst b Float where subst _ _ = id; substs _ = id
+instance Subst b Float where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
-instance Subst b Double where subst _ _ = id; substs _ = id
+instance Subst b Double where
+  subst _ _ = id
+  substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
 -- huh, apparently there's no instance Generic Integer.
 instance Subst b Integer where subst _ _ = id; substs _ = id
@@ -183,12 +225,17 @@ instance (Subst c a) => Subst c (Embed a)
 instance (Subst c e) => Subst c (Shift e) where
   subst x b (Shift e) = Shift (subst x b e)
   substs ss (Shift e) = Shift (substs ss e)
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
+
 
 instance (Subst c b, Subst c a, Alpha a, Alpha b) => Subst c (Bind a b) where
   subst x b (B p t) = B (subst x b p) (subst x b t)
   subst x a b = subst x a (forceBind b)
   substs ss (B p t) = B (substs ss p) (substs ss t)
   substs ss b = substs ss (forceBind b)
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}
 
 instance (Subst c p1, Subst c p2) => Subst c (Rebind p1 p2)
 
@@ -199,3 +246,5 @@ instance (Alpha p, Subst c p) => Subst c (TRec p)
 instance Subst a (Ignore b) where
   subst _ _ = id
   substs _ = id
+  {-# INLINE subst #-}
+  {-# INLINE substs #-}